legacy_apps/deepstream-retail-analytics/ds-retail-iva-frontend/requirements.txt
anyio==3.6.1
asgiref==3.5.2
backports.zoneinfo==0.2.1
charset-normalizer==2.1.0
cycler==0.11.0
distlib==0.3.4
Django==4.1
filelock==3.7.0
fonttools==4.34.4
h11==0.12.0
h2==4.1.0
hpack==4.0.0
httpcore==0.15.0
httpx==0.23.0
hyperframe==6.0.1
idna==3.3
kiwisolver==1.4.4
matplotlib==3.5.2
numpy==1.23.1
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
platformdirs==2.5.2
pykSQL==0.11.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
tqdm==4.64.0
urllib3==1.26.11
